saving for the day
Lots of progress in implementing the pseudo code algorithm. Have the whole structure in place but lacking some details. They are easy to find because pylint is telling me all about the unused variables. The real code looks alot like the pseudo making it easy to follow. Shortened some of the names.

Need to test the colllate acquisitions query. This is going to be crucial to test because it was just hand winged in there. Not even sure it is the best way to do the geographical search but so be it because running out of time.