fixes
footprint.convert() updated to generate a correct json object that osgeo can gutn into a Geometry. Found this problem and fix when it kept returning a None for the create.

footprint.coverage() updated because osgeo kept complaining that two areas did not overlap and thus no union. Bad process was to union all of the footprints into a mega area then intersect it with the AOI then compute that intersections area. Because some footprints may be detached and not have a union, intersect each bit with the AOI and compute its area which logically can be 0. Then sum them all for total coverage. Might be a little weird if they all cover the same place that can then tally more than the whole.

footprint.intersection_area() is a new function that computes the area of intersection between the AOI and a footprint.

footprint.project() updated to change from a zip object to an actual list. json.dumps() does not do this for us which causes problems.

footprint.union() deleted.

orbit.load() did not compute the local filename correctly. Fixed.

slc.load() updated to fix intersection problem. osgeo returns a None when there is no intersection rather than an object that has an Ares() of 0. Put in the additional checking.