correct coverage
footprint.coverage() summed all of the independent footprints. This does not account for the overlap of the footptints with themselves. Changed the algorithm to compute all of the largest independent (no overlap) footprints then intersect those with the AOI for an accurate measure of the coverage.

footprint.unionize() merges a list of footprints into the smallest set of independent (no intersection or overlap) footprints of largest size. It is recurisive and the python construct for/else is somewhat tricky because many do not use it.